+++ crates/zng-view/src/display_list.rs
@@ -681,10 +681,11 @@ fn display_item_to_webrender(
             image_id,
             image_size,
             rendering,
-            alpha_type,
+            image_size_hints,
             tile_size,
             tile_spacing,
         } => {
+            // !!: TODO try get alt downscaled image from cache
             images.push_display_list_img(
                 wr_list,
                 sc,
@@ -693,7 +694,6 @@ fn display_item_to_webrender(
                 *image_id,
                 *image_size,
                 *rendering,
-                *alpha_type,
                 *tile_size,
                 *tile_spacing,
             );