+++ crates/zng-wgt-image/src/node.rs
@@ -414,6 +414,7 @@ pub fn image_presenter() -> UiNode {
                                 render_tile_spacing,
                                 img,
                                 IMAGE_RENDERING_VAR.get(),
+                                vec![]
                             )
                         });
                     } else {
@@ -424,6 +425,7 @@ pub fn image_presenter() -> UiNode {
                             render_tile_spacing,
                             img,
                             IMAGE_RENDERING_VAR.get(),
+                            vec![]
                         );
                     }
                 }