+++ CHANGELOG.md
@@ -12,6 +12,10 @@
     - `Data` now holds `IpcBytes` directly.
     - Removed `Static` because it is always converted to `IpcBytes` anyway.
 
+* **Breaking** Refactor view-process API image display item.
+    - Removed `AlphaType` and relates items. Images are always premultiplied BGRA8.
+    - Display item can now hint future sizes for the image.
+
 * **Breaking** Refactor `zng::task::http` into a backend agnostic API.
     - Removed unmaintained `isahc` dependency.